Helpers

The Helpers sub-package contains both user interface function and functions useful for many of the internals such as data structure objects.

User interaction (@id User_interaction)

FaADE.Helpers.Boundary — Type

Boundary{BT<:BoundaryConditionType,BF<:Function,BN<:NodeType,BA<:Integer}

User defined boundary conditions for Dirichlet or Neumann. TODO: Robin boundaries

Constructors:

Boundary(type::BoundaryConditionType,RHS::Function,side::NodeType,axis::Int)
Boundary(type::BoundaryConditionType,RHS::Function,side::NodeType)

Inputs:

Dirichlet or Neumann (see BoundaryConditionType).

f(x,t) or f(x,y,t) depending on the function which gives the boundary condition.

Left, Right, Up, or Down (see NodeType)

1 if along the $x$ axis or 2 along the $y$ axis. TODO: Remove this requirement.

Returns:

Struct required for SAT construction.

FaADE.Helpers.PeriodicBoundary — Type

PeriodicBoundary(axis::Int)

User defined periodic boundary conditions for Periodic boundaries.

Inputs:

1 if periodic in $x$ and 2 if periodic in $y$.

Returns:

Struct required for SAT construction.

FaADE.Helpers.VariableCoefficientPDE1D — Type

VariableCoefficientPDE1D(u₀::Function,K::Function,order::Int,BCs::BoundaryConditionData...)

PDE Problem type for user input

Inputs:

Initial condition

Diffusion coefficient function

Solution order

Boundary conditions given by Boundary or PeriodicBoundary

Returns

Struct of data required for solver

FaADE.Helpers.VariableCoefficientPDE2D — Type

VariableCoefficientPDE2D

PDE Problem type for user input

Inputs:

Initial condition

Diffusion coefficient function in $x$

Diffusion coefficient function in $y$

Solution order

Boundary conditions given by Boundary or PeriodicBoundary

Returns

Struct of data required for solver

FaADE.Helpers.Grid1D — Type

Grid1D

Grid data structure for 1 dimensional problems.

`Grid1D{T}(𝒟::Vector,n::Integer)`

Inputs:

Vector of domain boundaries [x_0,x_N]

Number of grid points.

Returns:

Struct for 1D grid object containing vector of grid points, $\Delta x$ and $n$.

FaADE.Helpers.Grid2D — Type

Grid2D

Grid data structure for 2 dimensional problems.

`Grid2D{T}(𝒟x::Vector,𝒟y::Vector,nx::Integer,ny::Integer)`

Inputs:

Domain boundaries in $x$

Domain boundaries in $y$

Number of nodes in $x$

Number of nodes in $y$

Returns:

Struct for 2D grid object containing grid points in $x$ and $y$, $\Delta x$ and $\Delta y$, and $n_x$ and $n_y$.

Types

FaADE.Helpers.NodeType — Type

NodeType

Used by FirstDerivative() and SecondDerivative() to indicate which node type should be used.

Left and Right are used by the SAT() function to determine which side of the SAT to call.

Node types are:

Left

Internal

Right

Up

Down

FaADE.Helpers.BoundaryConditionType — Type

BoundaryCondition

Defines the boundary conditions, options are:

Dirichlet

Neumann

Robin

Periodic

Interface

Internal structures

These data blocks are used internally and are not interacted with by the user,

FaADE.Helpers.DataBlock — Type

DataBlock

Passed around internally between functions. Only contains data required for current timestep.

FaADE.Helpers.BoundaryData1D — Type

BoundaryData1D

Data structure for storage of SATs in 1 dimensional problems

FaADE.Helpers.BoundaryData2D — Type

BoundaryData2D

Data structure for storage of SATs in 2 dimensional problems

Moving data between blocks uses the following functions,

FaADE.Helpers.copyUtoSAT! — Function

copyUtoSAT!

Moves data from the solution u at a given boundary to the SAT_ field in BoundaryStorage structs. Or moves all data to SAT_ fields.

FaADE.Helpers.copySATtoU! — Function

copySATtoU!

Moves data from the SAT_ field on the given side in BoundaryStorage to u. Or moves all data from u to SAT_ fields.

FaADE.Helpers.addSATtoU! — Function

addSATtoU!

Add data from the SAT_ field on the given side in BoundaryStorage to u. Or add all data from u to SAT_ fields.

Nifty internal functions

FaADE.Helpers.check_order — Function

check_order

Ensure the users requested order makes sense.

FaADE.Helpers.check_boundary — Function

check_boundary

Some functions only work for Left or Right nodes.

FaADE.Helpers.halforder — Function

halforder

Returns $order/2$

FaADE.Helpers.BoundaryNodeInput — Function

BoundaryNodeInput

Returns the required number of nodes needed to compute FaADE.Derivatives.SecondDerivative on the boundary. $order == 2, return 1$, $order > 2, return 2order$

FaADE.Helpers.BoundaryNodeOutput — Function

BoundaryNodeOutput

Returns the size of the array output by FaADE.Derivatives.SecondDerivative on the boundary. $order == 2, return 1$, $order > 2, return order+order/2$

FaADE.Helpers.SATNodeOutput — Function

SATNodeOutput

Returns the number of nodes needed for the BoundaryData1D and BoundaryData2D data structures. $order == 2, return 2$, $order > 2, return order+order/2$

FaADE.Helpers.SelectLoopDirection — Function

SelectLoopDirection

Used to tell the SATs over which axis to loop

FaADE.Helpers.GetAxis — Function

GetAxis

Return the axis along which the node lies

FaADE.Helpers.GetDim — Function

GetDim

Return the data structures associated dimension

FaADE.Helpers.GetMinΔ — Function

GetMinΔ

Return the miniumum grid size between $\Delta x$ and $\Delta y$