Parallel map

The parallel operator for the 1D case is found here. In the 2D Cartesian case details can be found here.

FaADE.ParallelOperator.construct_grid — Method

construct_grid(χ::Function,grid::GridMultiBlock,z::Vector;interpmode=:nearest)

Constructs the backwards and forward planes for a multiblock grid. Returns a dictionary of PGrid objects corresponding to the grids in GridMultiBlock.

By default will return points for nearest neighbour interpolation.

FaADE.ParallelOperator.construct_grid — Method

construct_grid(χ::Function,grid::Grid2D,z::Vector)

Constructs the backwards and forward planes for a given plane

Inputs:

Field line ODE that returns $[x,y]$

GridType

z values of planes to trace to

Outputs:

ParallelGrid object (see ParallelGrid)

FaADE.ParallelOperator.ParallelData — Type

ParallelData{TT<:Real,
    DIM,
    TPARALLELGRID,
    TMAGNETICFIELD,
    TINTERPOLANT,
    TINTERCEPT} <: ParallelGridType

FaADE.ParallelOperator.ParallelData — Method

ParallelData(PGrid::ParallelGrid,G::Grid1D{TT},order::Int;κ=TT(1),intercept=nothing) where TT

FaADE.ParallelOperator.ParallelData — Method

ParallelData(PGrid::ParallelGrid,G::Grid2D{TT},order::Int;κ=TT(1),intercept=nothing,B=nothing,interpolant=nothing,remap=nothing) where {TT}

FaADE.ParallelOperator.ParallelMultiBlock — Type

ParallelMultiBlock{TT,DIM,IT}

Stores the parallel data for multiblock problems.

FaADE.ParallelOperator.applyParallelPenalty! — Function

applyParallelPenalty!

The default generated parallel penalty function for 2D (1D+parallel) problems

Operator in the 1D case is found here. In the 2D case details can be found here.

FaADE.ParallelOperator.applyParallelPenalty! — Method

Applies the parallel penalty for a multiblock problem

FaADE.ParallelOperator.applyParallelPenalty! — Method

2D Parallel penalty for single block problems

FaADE.ParallelOperator.applyParallelPenalty! — Method

1D apply parallel penalty

FaADE.ParallelOperator.compute_parallel_operator — Method

compute_parallel_operator

FaADE.ParallelOperator.computeglobalw! — Function

computeglobalw!

FaADE.ParallelOperator.computeglobalw! — Method

Compute the globalw for parallel blocks where a package is used for bicubic interpolation

FaADE.ParallelOperator.computeglobalw! — Method

Compute the globalw for parallel blocks where a custom interpolation scheme is used

Internals

FaADE.ParallelOperator.ParGrid — Type

ParGrid{TT,AT<:AbstractArray{TT}}

Storing the x-y coordinates of a parallel grid

FaADE.ParallelOperator.ParGridLinear — Type

ParGridLinear{TT,AT<:AbstractArray{TT}}

Storage for a parallel map using linear interpolation.

FaADE.ParallelOperator.ParallelGrid — Type

ParallelGrid{TT,DIM,AT,PA}

Stores the current, forward and backward planes for the parallel tracing.

In 2D arrays are of format $[(x_1,y_1),(x_1,y_2),...,(x_n,y_n)]$

FaADE.ParallelOperator.ParallelGridOptions — Type

Options for constructing the parallel grid

FaADE.ParallelOperator.ParallelGridType — Type

ParallelGridType

Abstract type for storage of

FaADE.ParallelOperator.ParallelMapType — Type

ParallelMapType

FaADE.ParallelOperator.readgrid — Method

readgrid(fname::String)

Read the ParallelGrid saved by savegrid into a ParallelGridType object.

FaADE.ParallelOperator.savegrid — Method

savegrid(fname,pargrid::ParallelGridType)

Saves the ParallelGrid object containing the backward and forward plane information to a jld2 file with path and filename specified by fname.