Simultaneous Approximation Terms

Users can create a generic SBP operator with the function,

FaADE.SATs.SAT — Function

SAT(BoundCond::BoundaryConditionData,grid::GridType,order::Int,solver)

Creates a Dirichlet, Neumann or Periodic SAT function(s).

Inputs:

BoundaryConditionData

GridType

Order of method

Solver type (see solvers)

Returns:

BD: SAT Struct

SATFn: SAT function(s)

If the solver is an implicit method, SATFn will have two methods, if explicit it will be a single method.

Note that this requires users to first create a boundary condition with boundary conditions

Boundary operators

The following are for boundary conditions

FaADE.SATs.SAT_Dirichlet — Type

SAT_Dirichlet

Storage of all objects needed for a Dirichlet SAT $\left. u\right|_{x_i} = g(t)$ where $i\in\{0,1\}$.

FaADE.SATs.SAT_Neumann — Type

SAT_Neumann

Simulatenous approximation term for Neumann boundary conditions $\left.\frac{\partial u}{\partial x}\right|_{x_i} = g(t)$ where $i\in\{0,N\}$

NodeType is either Left or Right

If periodic boundaries are applied

FaADE.SATs.SAT_Periodic — Type

SAT_Periodic

Storage of all objects needed for a Periodic SAT $u(x_0) = u(x_N)$ and $\left.\partial_x u\right|_{x_0} = \left.\partial_x u\right|_{x_N}$.

If a domain is split along some boundary then the following matches the values at the interface.

Boundary derivatives and penalties

The D_x^T B and B D_x operators are given by,

FaADE.SATs.BoundaryDerivativeTranspose — Function

BoundaryDerivativeTranspose

Used to construct EH⁻¹D₁ᵀ where E=E₀ or Eₙ.

FaADE.SATs.BoundaryDerivative — Function

BoundaryDerivative

Used to construct E₀D₁ or EₙD₁.

Penalties are constructed with the following,

FaADE.SATs.SATpenalties — Function

SATpenalties

Determines the penatly parameters for the given boundary conditions.

TODO: Remove inlineing

FaADE.SATs.hval — Function

hval(order::Int64)

Returns the value of $h^{-1}$ for the penalties

TODO: Remove inlineing