Time Integration

FaADE.solvers.solve — Function

solve

Solve the 1D or 2D PDE on the given grid.

Calls:

solve(Prob::VariableCoefficientPDE1D{T},grid::GridType{T,1},Δt::T,t_f::T,solver::Symbol;...)
solve(Prob::VariableCoefficientPDE2D{T},grid::GridType{T,2},Δt::T,t_f::T,solver::Symbol;...)

Inputs:

Prob: VariableCoefficientPDE1D or VariableCoefficientPDE2D

grid: Grid1D or Grid2D

Δt - the time step

t_f - the final time

solver::Symbol - must be :cgie currently.

Optional inputs:

adaptive::Bool=false true or false - Adaptive time stepping

source::Union{Nothing,Function}=nothing - Include a source term

penalty_func::Union{Nothing,Function}=nothing - penalty function for mapping points along field lines

TODO: Heavy optimisation required

FaADE.solvers.solution — Type

solution

Solution data structure, contains the initial condition and the solution at time of simulation termination. Also contains the grid structure and the PDE problem that was provided originally by the user.

Fields:

u, grid, Δt, t, problem, Δu

Integrators

Implicit

FaADE.solvers.conj_grad! — Function

conj_grad!

In-place conjugate gradient method.

See also build_H, A!, innerH

Explicit

Conjugate Gradient helpers

FaADE.Helpers.build_H — Function

build_H

Build the H matrix in the SBP operator D₁ = H⁻¹Q

FaADE.solvers.A! — Function

A!

Mutable $u - ΔtD(u)$

FaADE.Helpers.innerH — Type

innerH{T,N,VT<:AbstractVector{T}}

Method for a 1 or 2 dimensional H inner product.

Inputs:

GridType

Order of method

Returns:

Method-like struct for H-inner products