Summation by Parts operators

Below are the first and second derivative summation by parts operators.

First Derivative operators

The below operators are for first derivatives

FaADE.Derivatives.D₁ — Function

D₁

1D and 2D first derivative operator.

Also available as and internally uses in place operator D₁!.

D₁function D₁(u::AbstractVector{T},Δx::T;order::Integer=2)

1D implementation of $D_x$ operator.

D₁(u::AbstractMatrix{T},nx::Integer,ny::Integer,Δx::T,Δy::T;order::Integer=2)

2D implementation of $D_x$ operator

As iterators

FaADE.Derivatives.D₁! — Function

D₁!

1D and 2D in place first derivative operator.

See also FirstDerivativeBoundary! and FirstDerivativeInternal!.

D₁!(uₓ::AbstractVector{T},u::AbstractVector{T},n::Integer,Δx::T,order::Integer)

1D D₁!.

function D₁!(uₓ::AbstractArray{T},u::AbstractArray{T},n::Integer,Δ::T,order::Integer,dim::Integer)

1D implementation for 2D problems for D₁!.

Second Derivative operators

FaADE.Derivatives.D₂ — Function

D₂

D₂(u,c,Δx; order=2)

D₂(u,Δx,Δy,cx,cy; order=2)

1D and 2D second derivative operator.

Also available as and internally uses in place operator D₂!.

As iterators

The following functions are also available as iterators

FaADE.Derivatives.D₂! — Function

D₂!

1D and 2D in place second derivative operator.

Internally uses SecondDerivativeInternal and SecondDerivativeInternal.

Internal Functions

FaADE.Derivatives.FirstDerivative — Function

FirstDerivative

Allocating functions for first derivative, useful when need to add value to a matrix.

See also FirstDerivativeInternal! and FirstDerivativeBoundary

FirstDerivative(u::AbstractVector{T},Δx::T,::NodeType{:Internal},order::Integer)

Internal node function.

FirstDerivative(u::AbstractVector{T},Δx::T,node::NodeType,order::Integer)

Boundary node function.

FaADE.Derivatives.FirstDerivativeInternal — Function

FirstDerivativeInternal(u::AbstractVector{T},Δx::T,::NodeType{:Internal},order::Integer)

Single node 1D first derivative function.

See also FirstDerivativeInternal! for in place multi-node functions

FaADE.Derivatives.FirstDerivativeInternal! — Function

FirstDerivativeInternal!

1D and 2D in place functions for first derivative on internal nodes of a domain

FaADE.Derivatives.FirstDerivativeBoundary! — Function

FirstDerivativeBoundary!

1D in place function for first derivative on boundary nodes

FaADE.Derivatives.FirstDerivativeBoundaryStencil! — Function

FirstDerivativeBoundaryStencil!

2D first derivative on a boundary.

FirstDerivativeBoundaryStencil!(uₓ::AbstractArray{T},
    u::AbstractArray{T},
    xnode::NodeType,::NodeType{:Internal},
    Δx::T,Δy::T,nx::Integer,ny::Integer,order::Integer,
    Xrng::UnitRange=1:order,Yrng::UnitRange=1:order)

FirstDerivativeBoundaryStencil!(uₓ::AbstractArray{T},
    u::AbstractArray{T},
    ::NodeType{:Internal},ynode::NodeType,
    Δx::T,Δy::T,nx::Integer,ny::Integer,order::Integer,
    Xrng::UnitRange=1:order,Yrng::UnitRange=1:order)

FaADE.Derivatives.SecondDerivative — Function

SecondDerivative

Allocating functions for second derivative, useful when need to add value to a matrix.

See also SecondDerivativeInternal! and SecondDerivativeBoundary!